Add files via upload
In this project, I cleaned a bike purchases dataset as well as showcased some trends in the data that can help a bike company understand their customers better, and as a result, increase their bike sales. 

The project contains the raw data, the cleaned data, the pivot tables and the bike sales dashboard that was created using 3 different pivot tables.